training/requirements.txt
torch
python-chess
tqdm
numpy
dotenv
